refactor tests, add additional tests, extract FhirHelper from FhirConnector and fix spring use
