+++ src/test/java/de/numcodex/feasibility_gui_backend/query/broker/direct/DirectBrokerClientCqlTest.java
@@ -1,37 +1,88 @@
 package de.numcodex.feasibility_gui_backend.query.broker.direct;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static de.numcodex.feasibility_gui_backend.query.broker.direct.DirectBrokerClient.SITE_ID_LOCAL;
+import static de.numcodex.feasibility_gui_backend.query.collect.QueryStatus.COMPLETED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import de.numcodex.feasibility_gui_backend.query.QueryMediaType;
 import de.numcodex.feasibility_gui_backend.query.broker.QueryDefinitionNotFoundException;
 import de.numcodex.feasibility_gui_backend.query.broker.QueryNotFoundException;
 import de.numcodex.feasibility_gui_backend.query.broker.SiteNotFoundException;
+import de.numcodex.feasibility_gui_backend.query.collect.QueryStatusListener;
+import de.numcodex.feasibility_gui_backend.query.collect.QueryStatusUpdate;
 import java.io.IOException;
 import org.hl7.fhir.r4.model.Bundle;
+import org.hl7.fhir.r4.model.MeasureReport;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Answers;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class DirectBrokerClientCqlTest {
 
-    private static final Long TEST_BACKEND_QUERY_ID = 1L;
+    private static final Long TEST_BACKEND_QUERY_ID = 1234325L;
+    private static final String EXAMPLE_CQL = "example cql";
+    private static final int MEASURE_COUNT = 8723132;
 
     @SuppressWarnings("unused")
     @Mock(answer = Answers.RETURNS_DEEP_STUBS)
     FhirConnector fhirConnector;
 
-    @InjectMocks
+    @SuppressWarnings("unused")
+    @Mock(answer = Answers.RETURNS_DEEP_STUBS)
+    FhirHelper fhirHelper;
+
     DirectBrokerClientCql client;
+
+    @BeforeEach
+    void setUp() {
+        client = new DirectBrokerClientCql(fhirConnector, false, fhirHelper);
+    }
+
+    @Test
+    void testPublishExistingQuerySuccess() throws Exception {
+        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+        client.addQueryDefinition(queryId, QueryMediaType.CQL, EXAMPLE_CQL);
+        Bundle testBundle = new Bundle();
+        MeasureReport measureReport = new MeasureReport();
+        measureReport.addGroup().addPopulation().setCount(MEASURE_COUNT);
+        when(fhirHelper.createBundle(EXAMPLE_CQL, any(String.class), any(String.class))).thenReturn(testBundle);
+        when(fhirConnector.evaluateMeasure(any(String.class))).thenReturn(measureReport);
+
+        client.publishQuery(queryId);
+
+        verify(fhirConnector).transmitBundle(testBundle);
+        assertEquals(MEASURE_COUNT, client.findQuery(queryId).getResult());
+    }
+
+    @Test
+    void testPublishExistingQuerySuccessListener() throws Exception {
+        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+        client.addQueryDefinition(queryId, QueryMediaType.CQL, EXAMPLE_CQL);
+        var listener = mock(QueryStatusListener.class);
+        client.addQueryStatusListener(listener);
+        Bundle testBundle = new Bundle();
+        MeasureReport measureReport = new MeasureReport();
+        measureReport.addGroup().addPopulation().setCount(MEASURE_COUNT);
+        when(fhirHelper.createBundle(EXAMPLE_CQL, any(String.class), any(String.class))).thenReturn(testBundle);
+        when(fhirConnector.evaluateMeasure(any(String.class))).thenReturn(measureReport);
+
+        client.publishQuery(queryId);
+
+        verify(listener).onClientUpdate(TEST_BACKEND_QUERY_ID, new QueryStatusUpdate(client, queryId, SITE_ID_LOCAL, COMPLETED));
+    }
+
 
     @Test
     void testPublishNonExistingQuery() {
@@ -39,41 +90,42 @@ class DirectBrokerClientCqlTest {
     }
 
     @Test
-    void testPublishExistingQueryWithoutStructuredQueryDefinition() {
+    void testPublishExistingQueryWithoutQueryDefinition() {
         var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+
         assertThrows(QueryDefinitionNotFoundException.class, () -> client.publishQuery(queryId));
     }
 
     @Test
-    void testPublishExistingQueryWithIOExceptionInCreateBundle() throws IOException {
-        when(fhirConnector.createBundle(any(String.class), any(String.class), any(String.class))).thenThrow(IOException.class);
+    void testPublishExistingQueryWithIOExceptionInCreateBundle() throws Exception {
         var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
-        var cqlString = FhirConnector.getResourceFileAsString("gender-male.cql");
-        assertDoesNotThrow(() -> client.addQueryDefinition(queryId, QueryMediaType.CQL, cqlString));
+        client.addQueryDefinition(queryId, QueryMediaType.CQL, EXAMPLE_CQL);
+        when(fhirHelper.createBundle(eq(EXAMPLE_CQL), any(String.class), any(String.class))).thenThrow(IOException.class);
+
         assertThrows(IOException.class, () -> client.publishQuery(queryId));
     }
 
     @Test
-    void testPublishExistingQueryWithIOExceptionInTransmitBundle() throws IOException {
+    void testPublishExistingQueryWithIOExceptionInTransmitBundle() throws Exception {
+        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+        client.addQueryDefinition(queryId, QueryMediaType.CQL, EXAMPLE_CQL);
         doThrow(IOException.class).when(fhirConnector).transmitBundle(any(Bundle.class));
-        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
-        var cqlString = FhirConnector.getResourceFileAsString("gender-male.cql");
-        assertDoesNotThrow(() -> client.addQueryDefinition(queryId, QueryMediaType.CQL, cqlString));
+
         assertThrows(IOException.class, () -> client.publishQuery(queryId));
     }
 
     @Test
-    void testPublishExistingQueryWithIOExceptionInEvaluateMeasure() throws IOException {
+    void testPublishExistingQueryWithIOExceptionInEvaluateMeasure() throws Exception {
+        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+        client.addQueryDefinition(queryId, QueryMediaType.CQL, EXAMPLE_CQL);
         when(fhirConnector.evaluateMeasure(any(String.class))).thenThrow(IOException.class);
-        var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
-        var cqlString = FhirConnector.getResourceFileAsString("gender-male.cql");
-        assertDoesNotThrow(() -> client.addQueryDefinition(queryId, QueryMediaType.CQL, cqlString));
+
         assertThrows(IOException.class, () -> client.publishQuery(queryId));
     }
 
     @Test
-    void testGetSiteName() {
-        assertDoesNotThrow(() -> assertEquals("Local Server", client.getSiteName("1")));
+    void testGetSiteName() throws Exception {
+        assertEquals("Local Server", client.getSiteName("1"));
         assertThrows(SiteNotFoundException.class, () -> client.getSiteName("foo"));
         assertThrows(SiteNotFoundException.class, () -> client.getSiteName("something-else"));
         assertThrows(SiteNotFoundException.class, () -> client.getSiteName("FHIR Server"));
@@ -85,9 +137,12 @@ class DirectBrokerClientCqlTest {
     }
 
     @Test
-    void testCloseQuery() {
+    void testCloseQuery() throws Exception {
         var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
-        assertDoesNotThrow(() -> client.closeQuery(queryId));
+
+        client.closeQuery(queryId);
+
+        assertThrows( QueryNotFoundException.class, () -> client.findQuery(queryId));
     }
 
     @Test
@@ -98,6 +153,7 @@ class DirectBrokerClientCqlTest {
     @Test
     void testGetResultFeasibilityForUnknownSite() {
         var queryId = client.createQuery(TEST_BACKEND_QUERY_ID);
+
         assertThrows(SiteNotFoundException.class, () -> client.getResultFeasibility(queryId, "unknown-site-id"));
     }
